+++ Mobster/app/src/main/AndroidManifest.xml
@@ -81,7 +81,6 @@
         <activity
             android:name=".MapsActivity"
             android:label="@string/title_activity_maps" />
-        <activity android:name=".CommentActivity"></activity>
     </application>
 
 </manifest>
