+++ Mobster/app/src/main/java/Objects/Comment.java
@@ -6,20 +6,13 @@ package Objects;
 
 public class Comment {
     private String comment;
-    private String user;
 
-    public Comment(String comment, String user) {
+    public Comment(String comment) {
         this.comment = comment;
-        this.user = user;
     }
 
     public String getComment() {
         return comment;
     }
-
-    public String getUser() {
-        return user;
-    }
-
 
 }
