+++ Mobster/app/src/main/res/layout/comment_tab.xml
@@ -8,7 +8,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.the_great_amoeba.mobster.CommentActivity">
+    tools:context="com.the_great_amoeba.mobster.Results">
 
     <TextView
         android:text="No comments!"
@@ -23,7 +23,17 @@
     <ListView
         android:layout_width="match_parent"
         android:layout_height="match_parent"
+        android:id="@+id/comment_list"
+        android:layout_marginTop="63dp"
         android:layout_below="@+id/noCommentsText"
-        android:id="@+id/comment_list" />
+        android:layout_alignParentStart="true" />
+
+    <Button
+        android:text="POST"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:id="@+id/postButton"
+        android:layout_below="@+id/noCommentsText"
+        android:layout_centerHorizontal="true" />
 
 </RelativeLayout>
