+++ Mobster/app/src/main/res/layout/results_tab.xml
@@ -1,0 +1,34 @@
+<?xml version="1.0" encoding="utf-8"?>
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:orientation="vertical" android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:id="@+id/resultsLayout"
+    android:weightSum="1">
+
+    <TextView
+        android:text="Results"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:textSize="40sp"
+        android:gravity="center"
+        android:id="@+id/results_title" />
+
+    <ListView
+        android:layout_width="match_parent"
+        android:layout_height="119dp"
+        android:id="@+id/results_list"
+        android:layout_weight="0.16"
+        android:layout_alignParentBottom="true"
+        android:layout_alignParentStart="true"
+        android:layout_marginBottom="92dp" />
+
+    <com.github.mikephil.charting.charts.PieChart
+        android:id="@+id/chart"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_weight="0.78"
+        android:layout_alignParentStart="true"
+        android:layout_above="@+id/results_list"
+        android:layout_below="@+id/results_title" />
+
+</RelativeLayout>
