+++ Mobster/app/src/main/res/layout/vote_layout.xml
@@ -43,17 +43,17 @@
         android:layout_alignParentStart="true" />
 
     <EditText
+        android:id="@+id/commentText"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:inputType="textMultiLine"
-        android:lines="3"
-        android:gravity="top|left"
-        android:ems="10"
-        android:layout_below="@+id/commentLabel"
         android:layout_alignStart="@+id/commentLabel"
-        android:id="@+id/commentText"
+        android:layout_below="@+id/commentLabel"
+        android:layout_toStartOf="@+id/imageView_flag"
+        android:ems="10"
+        android:gravity="top|left"
         android:hint="Optional"
-        android:layout_toStartOf="@+id/imageView_flag" />
+        android:inputType="textPersonName"
+        android:lines="3" />
 
     <ScrollView
         android:id="@+id/scrollView1"
